Use MS install scripts to install the .NET SDKs
Use the Microsoft based install scripts to install the needed SDK. This removes the need for the SDK catalog updates. As is used now this will cause the most up to date .NET SDK to be installed. TODO: Call this script at least once per session so that the SDK can be updated, Or, possibly track to the Project's global.json file to make sure the required is installed.